backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
chromadb
pypdf
python-dotenv
tiktoken
langchain-huggingface
sentence-transformers
